Rename AllocProfilerData to AllocProfilerCommon, and move parts to source file
